backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==5.5
decorator==4.4.2
deprecation==2.1.0
exceptiongroup==1.2.1
eyed3==0.9.7
fastapi==0.110.2
filetype==1.2.0
h11==0.14.0
httptools==0.6.1
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mp3-tagger==1.0
numpy==1.26.2
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
python-decouple==3.8
python-dotenv==1.0.0
pytube==15.0.0
pytubemp3==0.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
starlette==0.37.2
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.11.0
urllib3==2.1.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
